requirements.txt
django<=2.0.9
martor
django-ratelimit